re-organization and renaming of core files
