+++ kolibri/core/assets/src/api-resource.js
