+++ kolibri/core/assets/src/api-resources/contentNode.js
@@ -1,4 +1,4 @@
-const Resource = require('../api_resource').Resource;
+const Resource = require('../api-resource').Resource;
 
 class ContentNodeResource extends Resource {
   constructor(...args) {
