+++ kolibri/core/assets/src/api-resources/facilityUser.js
@@ -1,4 +1,4 @@
-const Resource = require('../api_resource').Resource;
+const Resource = require('../api-resource').Resource;
 const logging = require('logging').getLogger(__filename);
 
 class FacilityUserResource extends Resource {
