+++ kolibri/core/assets/src/core-app/asset-loader.js
