+++ kolibri/core/assets/src/core-app/index.js
@@ -1,8 +1,8 @@
 
 // include global styles
 require('normalize.css');
-require('./font-NotoSans.css');
-require('./core-global.styl');
+require('../styles/font-NotoSans.css');
+require('../styles/core-global.styl');
 
 // polyfill for older browsers
 require('babel-polyfill');
@@ -16,5 +16,5 @@ logging.setDefaultLevel(process.env.NODE_ENV === 'production' ? 2 : 0);
 //
 // TODO - clarify how setting the `src_file` attribute in the core
 // kolibri_plugin.py attaches this object to the global variable.
-const CoreAppConstructor = require('./core_app_constructor');
+const CoreAppConstructor = require('./constructor');
 module.exports = new CoreAppConstructor();
