+++ kolibri/core/assets/src/styles/core-global.styl
@@ -17,7 +17,7 @@
 
 
 
-@require '~core-theme.styl'
+@require 'core-theme.styl'
 
 
 // http://www.paulirish.com/2012/box-sizing-border-box-ftw/
