+++ kolibri/core/assets/src/vue/error-box/index.vue
